SKALE-1859 update tests to match new architecture
