+++ tests/constants.py
@@ -8,6 +8,7 @@ DIR_PATH = os.path.dirname(os.path.realpath(__file__))
 DEFAULT_NODE_NAME = 'test_node'
 SECOND_NODE_NAME = 'test_node_2'
 DEFAULT_NODE_HASH = '23bdf46c41fa300e431425baff124dc31625b34ec09b829f61aa16ab0102ca8d'
+DEFAULT_NODE_PORT = 3000
 
 TEST_CONTRACT_NAME = 'NodesFunctionality'
 TEST_CONTRACT_NAME_HASH = 'f88bdb637038c4be9f72381c0db0b0d7b7f369cfdd49619ee7e48aa7940482b9'
